Исправлена ошибка при логине только что созданного пользователя

See merge request kurator_skillbox/python_django_team33!45

+++ src/market/profiles/views.py
@@ -58,9 +58,10 @@ class LoginView(LoginView):
         response = super().post(request, *args, **kwargs)
         if self.request.user.is_authenticated:
             user = User.objects.get(pk=self.request.user.pk)
-            user_cart = UserCart.objects.get(user=user)
-            rebuild_cart(request.session, user_cart.items)
-            user_cart.delete()
+            user_cart = UserCart.objects.filter(user=user).first()
+            if user_cart:
+                rebuild_cart(request.session, user_cart.items)
+                user_cart.delete()
         return response
 
 class LogoutView(LogoutView):
